refactor: tighten lifecycle runtime topology
