+++ scripts/layering/runtime-command-cutover-extensions.ts
@@ -1,6 +1,15 @@
 import { parseSync } from 'oxc-parser';
 import { propertyName, visitAst, type ProductionSource } from './cutover-policy-ast.ts';
+import { lineOf } from './runtime-command-cutover-ast.ts';
 import type { UnruledViolation } from './runtime-command-cutover-model.ts';
+
+export {
+  applicationLifecycleDurableResourceViolations,
+  closeLifecycleRouteBindingViolations,
+  openLifecycleRouteBindingViolations,
+  prepareLifecycleRouteBindingViolations,
+  runtimeLifecycleRouteBindingViolations,
+} from './runtime-command-cutover-lifecycle.ts';
 
 type AstNode = Record<string, unknown>;
 
@@ -20,259 +29,6 @@ const APP_STATE_LEGACY_IMPORT_SOURCES = new Set([
   '../../platforms/harmonyos/app-lifecycle.ts',
 ]);
 const APP_STATE_LEGACY_CALLS = new Set(['getAndroidAppState', 'getHarmonyAppState']);
-
-/**
- * `admitRuntimeUse` performs the facts inspection and the bind itself; `admitRuntimeOperations`
- * performs the inspection and hands back the bind for a route that selects among exactly-typed
- * uses. Both live in the one admission module, which this file gates separately.
- */
-const RUNTIME_ADMISSION_MODULE = 'src/daemon/runtime-admission.ts';
-const APPLICATION_RESOURCES_FILE = 'src/platform-runtime-application-resources.ts';
-const RUNTIME_ADMISSION_HELPERS = ['admitRuntimeUse', 'admitRuntimeOperations'];
-
-type LifecycleAdmissionRule = Readonly<{ functionName: string }>;
-
-type LifecycleHandlerRule = Readonly<{
-  command: 'open' | 'prepare' | 'close' | 'runtime';
-  file: string;
-  admissions: readonly LifecycleAdmissionRule[];
-  forbiddenCalls: readonly string[];
-}>;
-
-// A handler that calls its injected gateway directly would bypass the shared admission seam, so
-// the raw ports are forbidden call names in every lifecycle route.
-const RAW_RUNTIME_GATEWAY_CALLS = ['inspectFacts', 'bindDevice'];
-
-const LIFECYCLE_HANDLER_RULES = {
-  open: {
-    command: 'open',
-    file: 'src/daemon/handlers/session-open.ts',
-    admissions: [{ functionName: 'admitOpenRuntime' }],
-    forbiddenCalls: [
-      ...RAW_RUNTIME_GATEWAY_CALLS,
-      'dispatchCommand',
-      'resolveSoleForegroundIosApp',
-      'runAndroidAdb',
-      'runXcrun',
-      'shutdownDeviceTarget',
-    ],
-  },
-  prepare: {
-    command: 'prepare',
-    file: 'src/daemon/handlers/session-prepare.ts',
-    admissions: [{ functionName: 'admitPrepareRuntime' }],
-    forbiddenCalls: [
-      ...RAW_RUNTIME_GATEWAY_CALLS,
-      'dispatchCommand',
-      'prepareIosRunner',
-      'runXcrun',
-      'shutdownDeviceTarget',
-    ],
-  },
-  close: {
-    command: 'close',
-    file: 'src/daemon/handlers/session-close-runtime-admission.ts',
-    admissions: [{ functionName: 'admitCloseRuntime' }],
-    forbiddenCalls: [
-      ...RAW_RUNTIME_GATEWAY_CALLS,
-      'dispatchCommand',
-      'shutdownDeviceTarget',
-      'runAndroidAdb',
-      'runXcrun',
-    ],
-  },
-  runtime: {
-    command: 'runtime',
-    file: 'src/daemon/handlers/session-runtime-command.ts',
-    admissions: [
-      { functionName: 'admitClearRuntime' },
-      { functionName: 'admitPortReverseRuntime' },
-    ],
-    forbiddenCalls: [
-      ...RAW_RUNTIME_GATEWAY_CALLS,
-      'dispatchCommand',
-      'shutdownDeviceTarget',
-      'runAndroidAdb',
-      'runXcrun',
-    ],
-  },
-} satisfies Record<'open' | 'prepare' | 'close' | 'runtime', LifecycleHandlerRule>;
-
-/**
- * Lifecycle command routes own exactly one side-effect-free facts admission and one binding
- * gateway per native plan. The `runtime` descriptor has two disjoint native plans (clear and
- * provider port-reverse); each is checked independently so session-only set/show policy stays out
- * of the platform-runtime denominator.
- */
-export const openLifecycleRouteBindingViolations = (sources: ReadonlyMap<string, string>) =>
-  lifecycleHandlerRouteBindingViolations(sources, LIFECYCLE_HANDLER_RULES.open);
-export const prepareLifecycleRouteBindingViolations = (sources: ReadonlyMap<string, string>) =>
-  lifecycleHandlerRouteBindingViolations(sources, LIFECYCLE_HANDLER_RULES.prepare);
-export const closeLifecycleRouteBindingViolations = (sources: ReadonlyMap<string, string>) =>
-  lifecycleHandlerRouteBindingViolations(sources, LIFECYCLE_HANDLER_RULES.close);
-export const runtimeLifecycleRouteBindingViolations = (sources: ReadonlyMap<string, string>) =>
-  lifecycleHandlerRouteBindingViolations(sources, LIFECYCLE_HANDLER_RULES.runtime);
-
-function lifecycleHandlerRouteBindingViolations(
-  sources: ReadonlyMap<string, string>,
-  rule: LifecycleHandlerRule,
-): UnruledViolation[] {
-  const source = sources.get(rule.file);
-  if (source === undefined) {
-    return [
-      {
-        file: rule.file,
-        line: 1,
-        message: `${rule.command} lifecycle handler module is missing`,
-      },
-    ];
-  }
-  const program = parseSync(rule.file, source).program as AstNode;
-  const violations: UnruledViolation[] = [];
-  for (const admission of rule.admissions) {
-    const functionNode = namedFunction(program, admission.functionName);
-    if (functionNode === undefined) {
-      violations.push({
-        file: rule.file,
-        line: 1,
-        message: `${rule.command} must retain its ${admission.functionName} admission seam`,
-      });
-      continue;
-    }
-    const admissionCalls = RUNTIME_ADMISSION_HELPERS.reduce(
-      (total, helper) => total + countNamedCalls(functionNode, helper),
-      0,
-    );
-    if (admissionCalls !== 1) {
-      violations.push({
-        file: rule.file,
-        line: lineOf(source, functionNode),
-        message: `${rule.command} ${admission.functionName} must make one shared runtime admission call (found ${admissionCalls})`,
-      });
-    }
-  }
-  visitAst(program, (node) => {
-    const call = callName(node);
-    if (call === undefined || !rule.forbiddenCalls.includes(call)) return;
-    violations.push({
-      file: rule.file,
-      line: lineOf(source, node),
-      message: `${rule.command} handler reaches legacy or local platform call ${call} outside its bound runtime`,
-    });
-  });
-  violations.push(...sharedRuntimeAdmissionViolations(sources));
-  return violations;
-}
-
-/**
- * Every lifecycle row delegates its "one side-effect-free inspection, one bind" promise to the
- * shared module, so the promise is proven once at that module rather than re-detected per handler.
- */
-function sharedRuntimeAdmissionViolations(
-  sources: ReadonlyMap<string, string>,
-): UnruledViolation[] {
-  const source = sources.get(RUNTIME_ADMISSION_MODULE);
-  if (source === undefined) {
-    return [
-      {
-        file: RUNTIME_ADMISSION_MODULE,
-        line: 1,
-        message: 'shared runtime admission module is missing',
-      },
-    ];
-  }
-  const program = parseSync(RUNTIME_ADMISSION_MODULE, source).program as AstNode;
-  const admit = namedFunction(program, 'admitRuntimeOperations');
-  if (admit === undefined) {
-    return [
-      {
-        file: RUNTIME_ADMISSION_MODULE,
-        line: 1,
-        message: 'shared runtime admission must expose admitRuntimeOperations',
-      },
-    ];
-  }
-  const violations: UnruledViolation[] = [];
-  for (const [helper, role] of [
-    ['requireFactsInspection', 'facts inspection'],
-    ['requireDeviceBinding', 'binding'],
-  ] as const) {
-    const calls = countNamedCalls(admit, helper);
-    if (calls !== 1) {
-      violations.push({
-        file: RUNTIME_ADMISSION_MODULE,
-        line: lineOf(source, admit),
-        message: `shared runtime admission must make one ${role} call (found ${calls})`,
-      });
-    }
-  }
-  return violations;
-}
-
-/** Durable lifecycle evidence that remains shared by the four descriptor rows. */
-export function applicationLifecycleDurableResourceViolations(
-  sources: ReadonlyMap<string, string>,
-): UnruledViolation[] {
-  const violations: UnruledViolation[] = [];
-  // The neutral gateway registers the fence; the host composition names the durable owners. Both
-  // halves are required, and neither may absorb the other's role.
-  const gateway = sources.get('src/platform-runtime-gateway.ts');
-  if (
-    gateway === undefined ||
-    !gateway.includes('runStartupRecoveryFence') ||
-    gateway.includes('androidApplications') ||
-    gateway.includes('appleApplications')
-  ) {
-    violations.push({
-      file: 'src/platform-runtime-gateway.ts',
-      line: 1,
-      message:
-        'the composed runtime gateway must register the startup fence and name no platform-specific durable owner',
-    });
-  }
-  const resources = sources.get(APPLICATION_RESOURCES_FILE);
-  if (
-    resources === undefined ||
-    resources.indexOf('hasTestImeRecoveryEvidence') < 0 ||
-    resources.indexOf('hasTestImeRecoveryEvidence') > resources.indexOf('recoverTestImeStartup')
-  ) {
-    violations.push({
-      file: APPLICATION_RESOURCES_FILE,
-      line: 1,
-      message:
-        'durable startup recovery must gate lazy Android test-IME recovery behind its marker evidence',
-    });
-  }
-  const ime = sources.get('src/platforms/android/ime-lifecycle.ts');
-  if (
-    ime === undefined ||
-    ime.indexOf('await waitForStartupRecoveryFence') < 0 ||
-    ime.indexOf('await waitForStartupRecoveryFence') >
-      ime.indexOf('const adb = resolveAndroidAdbExecutor')
-  ) {
-    violations.push({
-      file: 'src/platforms/android/ime-lifecycle.ts',
-      line: 1,
-      message:
-        'Android test-IME mutation must wait for startup recovery before loading ADB mechanics',
-    });
-  }
-  const captureKitUnavailable = sources.get(
-    'packages/capture-kit/src/platform-runtime-unavailable.ts',
-  );
-  const captureKitIndex = sources.get('packages/capture-kit/src/index.ts');
-  if (
-    captureKitUnavailable !== undefined ||
-    captureKitIndex?.includes('createUnavailablePlatformRuntime') === true
-  ) {
-    violations.push({
-      file: 'packages/capture-kit/src/index.ts',
-      line: 1,
-      message: 'capture-kit must not own a generic platform-runtime lifecycle role',
-    });
-  }
-  return violations;
-}
 
 /**
  * `devices` proves its single inventory route by binding identity, not by name: the
@@ -540,33 +296,4 @@ function identifierName(node: unknown): string | undefined {
   return record['type'] === 'Identifier' && typeof record['name'] === 'string'
     ? record['name']
     : undefined;
-}
-
-function namedFunction(program: AstNode, expected: string): AstNode | undefined {
-  let found: AstNode | undefined;
-  visitAst(program, (node) => {
-    if (found !== undefined || node['type'] !== 'FunctionDeclaration') return;
-    if (identifierName(node['id']) === expected) found = node;
-  });
-  return found;
-}
-
-function countNamedCalls(functionNode: AstNode, expected: string): number {
-  let count = 0;
-  visitAst(functionNode['body'], (node) => {
-    if (callName(node) === expected) count += 1;
-  });
-  return count;
-}
-
-function callName(node: AstNode): string | undefined {
-  if (node['type'] !== 'CallExpression') return undefined;
-  const callee = node['callee'] as AstNode | undefined;
-  if (callee?.['type'] === 'Identifier') return identifierName(callee);
-  if (callee?.['type'] === 'MemberExpression') return propertyName(callee['property']);
-  return undefined;
-}
-
-function lineOf(source: string, node: AstNode): number {
-  return source.slice(0, (node['start'] as number | undefined) ?? 0).split('\n').length;
 }
