+++ src/platform-runtime-gateway.test.ts
@@ -1,56 +1,23 @@
-import type { ProviderDeviceRuntime } from '@agent-device/contracts/device';
-import type { Interactor } from '@agent-device/contracts/interaction';
-import type {
-  ApplicationLifecycleRuntimeOperations,
-  DeviceBinding,
-  PlatformRuntimeHost,
-  PlatformRuntimeOperations,
-  PlatformRuntimeOwner,
-  PlatformRequestScope,
-  RuntimeFacts,
-  RuntimeOwnerRef,
-} from '@agent-device/contracts/platform';
-import {
-  applicationLifecycleOperationFacts,
-  availableApplicationLifecycleOperations,
-  providerRuntimeOwner,
-} from '@agent-device/contracts/platform';
+import type { PlatformRuntimeHost, PlatformRuntimeOwner } from '@agent-device/contracts/platform';
+import { providerRuntimeOwner } from '@agent-device/contracts/platform';
 import type { DeviceInfo } from '@agent-device/kernel/device';
-import { createLimrunRuntime, type LimrunRuntimeDependencies } from '@agent-device/provider-limrun';
+import { createLimrunRuntime } from '@agent-device/provider-limrun';
 import { describe, expect, test, vi } from 'vitest';
-import { unavailableDeploymentAndShutdownOperationFacts } from './__tests__/test-utils/runtime-operation-facts.ts';
 import { handleSessionStateCommands } from './daemon/handlers/session-state.ts';
 import { createRequestRuntimeBindings } from './daemon/request-runtime-binding.ts';
 import { makeSessionStore } from './__tests__/test-utils/store-factory.ts';
 import { withTestDeviceInventory } from './__tests__/test-utils/device-inventory-gateways.ts';
+import { createComposedPlatformRuntimeGateway } from './platform-runtime-gateway.ts';
 import {
-  createComposedPlatformRuntimeGateway,
-  type PlatformRuntimeProviderRegistration,
-} from './platform-runtime-gateway.ts';
-
-const device: DeviceInfo = {
-  platform: 'apple',
-  appleOs: 'ios',
-  id: 'limrun:ios:lease-a',
-  name: 'Provider iOS',
-  kind: 'simulator',
-  target: 'mobile',
-  booted: true,
-};
-const scope: PlatformRequestScope = {
-  signal: new AbortController().signal,
-  diagnostics: { emit: () => {} },
-  progress: { report: () => {} },
-};
-
-const LIFECYCLE_FACETS = [
-  ['openTarget', ['resolveOpenTarget', 'prepareApplicationOpen', 'openApplication']],
-  ['prepareAppleRunner', ['prepareAppleRunner']],
-  ['closeTarget', ['closeApplication', 'finalizeApplicationClose']],
-  ['runtimeHints', ['applyRuntimeHints', 'clearRuntimeHints']],
-] as const;
-
-type LifecycleFacet = (typeof LIFECYCLE_FACETS)[number][0];
+  gatewayFixture as gateway,
+  gatewayFixtureDevice as device,
+  gatewayFixtureScope as scope,
+  LIFECYCLE_FACETS,
+  limrunTestDependencies,
+  providerLifecycleOwnerFixture as providerLifecycleOwner,
+  providerRuntimeFixture as providerRuntime,
+  runtimeOwnerFixture as runtimeOwner,
+} from './platform-runtime-gateway.fixtures.ts';
 
 describe('composed platform runtime gateway', () => {
   // Which resources need recovering is the host's composition (see
@@ -389,213 +356,3 @@ describe('composed platform runtime gateway', () => {
     },
   );
 });
-
-function gateway(registrations: readonly PlatformRuntimeProviderRegistration[]) {
-  return createComposedPlatformRuntimeGateway({
-    modules: new Map(),
-    loadHost: async () => ({}) as PlatformRuntimeHost,
-    providerRuntimes: registrations.map(({ runtime }) => runtime),
-    providerModules: registrations,
-  });
-}
-
-function providerRuntime(options: {
-  ref: RuntimeOwnerRef;
-  provider?: string;
-  ownsDevice?: (device: DeviceInfo) => boolean;
-  mismatch?: 'owner' | 'device' | 'facts';
-  disposed?: () => Promise<void>;
-  load?: () => Promise<PlatformRuntimeOwner>;
-}): PlatformRuntimeProviderRegistration {
-  const owner = runtimeOwner(options);
-  const runtime: ProviderDeviceRuntime = {
-    provider: options.provider ?? 'limrun',
-    leaseLifecycle: {},
-    deviceInventoryProvider: async () => null,
-    ownsDevice: options.ownsDevice ?? (() => true),
-    getInteractor: () => undefined,
-    shutdown: async () => {},
-  };
-  return {
-    runtime,
-    module: {
-      owner: options.ref as Extract<RuntimeOwnerRef, { kind: 'provider-runtime' }>,
-      loadRuntime: options.load ?? (async () => owner),
-    },
-  };
-}
-
-function runtimeOwner(options: {
-  ref: RuntimeOwnerRef;
-  mismatch?: 'owner' | 'device' | 'facts';
-  providerMode?: 'local' | 'transport-composed' | 'provider-runtime';
-  disposed?: () => Promise<void>;
-}): PlatformRuntimeOwner {
-  return {
-    owner: options.ref,
-    ownsDevice: () => true,
-    inspectFacts: async () => binding(options).facts,
-    bind: async () => binding(options),
-    shutdown: async () => {},
-  };
-}
-
-function binding(options: {
-  ref: RuntimeOwnerRef;
-  mismatch?: 'owner' | 'device' | 'facts';
-  providerMode?: 'local' | 'transport-composed' | 'provider-runtime';
-  disposed?: () => Promise<void>;
-}): DeviceBinding<PlatformRuntimeOperations> {
-  const bindingDevice = options.mismatch === 'device' ? { ...device, id: 'wrong' } : device;
-  const bindingOwner =
-    options.mismatch === 'owner' ? providerRuntimeOwner('limrun', 'wrong') : options.ref;
-  return {
-    device: bindingDevice,
-    owner: bindingOwner,
-    facts: {
-      device: {
-        family: bindingDevice.platform,
-        appleOs: bindingDevice.appleOs,
-        kind: bindingDevice.kind,
-        target: bindingDevice.target,
-        providerMode:
-          options.mismatch === 'facts'
-            ? 'transport-composed'
-            : (options.providerMode ?? 'provider-runtime'),
-      },
-      operations: unavailableFacts(),
-    },
-    operations: {},
-    [Symbol.asyncDispose]: options.disposed ?? (async () => {}),
-  };
-}
-
-function unavailableFacts() {
-  const unavailable = { available: false, reason: 'unsupported-provider-mode' } as const;
-  return {
-    ...unavailableDeploymentAndShutdownOperationFacts,
-    appLogInspect: unavailable,
-    appLogDoctor: unavailable,
-    appLogStart: unavailable,
-    appLogReattach: unavailable,
-    appLogCleanup: unavailable,
-    appState: unavailable,
-    networkDump: unavailable,
-    screenRecordingStart: unavailable,
-    screenRecordingReattach: unavailable,
-    screenRecordingCleanup: unavailable,
-    ensureReady: unavailable,
-    bootTarget: unavailable,
-    bootTargetHeadless: unavailable,
-    listApps: unavailable,
-    ...applicationLifecycleOperationFacts({
-      resolveOpenTarget: unavailable,
-      prepareApplicationOpen: unavailable,
-      openApplication: unavailable,
-      applyRuntimeHints: unavailable,
-      clearRuntimeHints: unavailable,
-      closeApplication: unavailable,
-      finalizeApplicationClose: unavailable,
-      prepareAppleRunner: unavailable,
-      configureProviderPortReverse: unavailable,
-    }),
-  };
-}
-
-function providerLifecycleOwner(
-  ref: Extract<RuntimeOwnerRef, { kind: 'provider-runtime' }>,
-  unavailableFacet: LifecycleFacet,
-): PlatformRuntimeOwner {
-  const unavailable = Object.freeze({
-    available: false,
-    reason: 'unsupported-provider-mode',
-  } as const);
-  const available = Object.freeze({ available: true } as const);
-  const lifecycleFacts = applicationLifecycleOperationFacts({
-    resolveOpenTarget: unavailableFacet === 'openTarget' ? unavailable : available,
-    prepareApplicationOpen: unavailableFacet === 'openTarget' ? unavailable : available,
-    openApplication: unavailableFacet === 'openTarget' ? unavailable : available,
-    applyRuntimeHints: unavailableFacet === 'runtimeHints' ? unavailable : available,
-    clearRuntimeHints: unavailableFacet === 'runtimeHints' ? unavailable : available,
-    closeApplication: unavailableFacet === 'closeTarget' ? unavailable : available,
-    finalizeApplicationClose: unavailableFacet === 'closeTarget' ? unavailable : available,
-    prepareAppleRunner: unavailableFacet === 'prepareAppleRunner' ? unavailable : available,
-    configureProviderPortReverse: unavailable,
-  });
-  const facts: RuntimeFacts<PlatformRuntimeOperations> = {
-    device: {
-      family: device.platform,
-      appleOs: device.appleOs,
-      kind: device.kind,
-      target: device.target,
-      providerMode: 'provider-runtime',
-    },
-    operations: { ...unavailableFacts(), ...lifecycleFacts },
-  };
-  const operations = availableApplicationLifecycleOperations(
-    lifecycleOperations(),
-    facts.operations,
-  );
-  return {
-    owner: ref,
-    ownsDevice: () => true,
-    inspectFacts: async () => facts,
-    bind: async () => ({
-      device,
-      owner: ref,
-      facts,
-      operations: operations as PlatformRuntimeOperations,
-      [Symbol.asyncDispose]: async () => {},
-    }),
-    shutdown: async () => {},
-  };
-}
-
-function lifecycleOperations(): ApplicationLifecycleRuntimeOperations {
-  return {
-    resolveOpenTarget: async () => ({}),
-    prepareApplicationOpen: async () => {},
-    openApplication: async () => ({ timing: {} }),
-    applyRuntimeHints: async () => {},
-    clearRuntimeHints: async () => {},
-    closeApplication: async () => {},
-    finalizeApplicationClose: async () => {},
-    prepareAppleRunner: async () => ({ runner: {}, connectMs: 0, healthCheckMs: 0 }),
-    configureProviderPortReverse: async () => undefined,
-  };
-}
-
-const limrunTestDependencies = {
-  clientVersion: 'test',
-  android: {
-    createInteractor: () => ({}) as Interactor,
-    createPortReverse: async () => ({
-      ensure: async () => {},
-      remove: async () => {},
-      removeAllOwned: async () => {},
-    }),
-    inferAppName: async () => 'unused',
-    listApps: async () => [],
-    getForegroundApp: async () => undefined,
-    getKeyboardState: async () => ({ visible: false, inputOwner: 'unknown' as const }),
-    dismissKeyboard: async () => ({
-      visible: false,
-      inputOwner: 'unknown' as const,
-      attempts: 0,
-      wasVisible: false,
-      dismissed: false,
-    }),
-    readLogs: async () => '',
-    adbError: async () => {
-      throw new Error('unused');
-    },
-  },
-  host: {
-    runAdb: async () => ({ stdout: '', stderr: '', exitCode: 0 }),
-    archiveDirectory: async () => {},
-  },
-  ios: {
-    resolveAppAlias: async (app: string) => app,
-    readBundleAppName: async () => undefined,
-  },
-} satisfies LimrunRuntimeDependencies;
